Set the height and width of the layout.html screen from variables
Add a pair of template variables (screen_height and screen_width)
that can be used to set the height and width of the layout.html screen.
Default these to 1080px and 1920px to produce a standard SmartScreen
display if they are omitted. Take note that these are substituted
as CSS lengths and so need to be qualified with CSS length units.

Setting them to '100%' produces a responsive grid (though currently
widgets aren't particularly responsive so the usefulness of this is
questionable). However being able to set height to '100%' requires
the associated change to smartpanel.css.