Rework to only support a fixed aspect ratio
layout.html now takes a single optional variable `screen_width` which
should contain a CSS length measurement, defaulting to 1920.
Heights are calculated based on this and a 16:9 aspect ratio.
screen_width must not be in percent, but almost anything else
that results in a CSS measurement should work. In particular you
can use 100vw for 'full display width'.