Re-re-work grid-container size to what we hope is now correct
We want the <div class="grid-container ..."> to have

    width={{display_width}}

and

    height=calc(({{display_width}} * 0.5625) - 60px)

Given that {{display_width}} * 0.5625 gives us the 16:9 display_height
that corresponds to display_width. Setting (or defaulting) display_width
to 1920px (or almost any other CSS length) gives a fixed-width version,
setting it to 100vw will give a responsive one. However you can't set
it to 100% (because that gets interpreted differently when setting
height from when setting width)

This change also moves from calling the parameter screen_width to
calling it display_width.