Add class='widget' to container element; add trailing '/' to static_url
Adding class='widget' to thwe widget container element makes it possible
to target CSS rules for all widgets without disturbing the rest of the
page.

In the jw35/acp_loby_screen framework the static_url passed to widgets
includes a trailing '/'. This makes concatenation easier and is included
here for that reason and compatibility.

Also updated widgets/README.md to match.