Apply JS escaping to transferred data (and fix calling sequence)
JavaScript escape the JSON representation of the widget configuration
data when embedding it in HTML pages and use JSON.parse to de-serialize
in JavaScript. This reduces the possibility of rogue data breaking out
either into the HTML interpretation context or into JavaScript
execution context.

See
https://www.owasp.org/index.php/XSS_(Cross_Site_Scripting)_Prevention_Cheat_Sheet#RULE_.233_-_JavaScript_Escape_Before_Inserting_Untrusted_Data_into_JavaScript_Data_Values
for a wider discussion.

[[At the same time, fix the widget calling convention in
layout_config.html and layout_config2.html to match the (correct)
calls in layout.html]]