Add support for displaying "enum" values as human-readable strings
Add a brutusin-json-forms extension to the layout_config.html template
to recognise a "display" key in a Json schema and use its value as a
map to display human-readable strings in place of raw enum values in
form select fields.

Update station_board, stop_timetable and weather schemas to make use
of this, and extend the number of stations and forecast locations
supported.

Add cached copies of the documents defining station codes and forecast
locations.