Add DEBUG global and catch up other changes
Widgets can assume the existence of a "DEBUG" global which can contain
their name if they are to emit debug logs to consle.log() (see widget
spec under 'Widget JavaScript objects'). This global needs to be created
by the framework before instantiating widgets and this patch does so.

Also apply some other changes that have already been made to layout.html
and layout_config.html but never made it here. This may well be
unnecessary, but as long as this file is in the repository...

Closes #97