Add additional widget styling to get ul/ol/li working
smartpannel.css redefines many ul/ol/li CSS properties for its
own purposes. This edit attempts to reset them to something like
default for the benefit of widgets (in particular message_area).
Values are based on Chrome defaults.

This is to some extent a stop-gap pending review of CSS usage
throughout the framework.

Closes #98