Further isolate CSS rules in widgets.css
widgets.css was unconditionally redefining h1, h2 and table which was
problematic for display of non-widget information on pages that included
widgets.css. Isolate those definitions so they only apply within
objects with the .widget class.

Further-further isolate the table definitions, which apply only
to timetables, to a .timetables class and add this as needed to
station_board.html and stop_timetable.js.

Add a knock-on change to smartpanel.css to re-establish
the font definition on the logo/credit header which was previously
'leaking' from widgets.css.