Alter util.get_config() to match on multiple fields
Physical zones only appear in sys/data_zone_config/list.json once,
so to find the configuration for a particular zone it's necessary to
look for entries with matching zone.id or zone.reverse.id.

This patch alters util.get_config() to take a list of field names
to match against, and alters the get_*_config() methods in the
aq, parking and zone views to pass a suitable list. Only the zone
view passes a list of length other than one.