Move /transport/api endpoints to /api/v1/transport
For consistency with the developing /api/ endpoint, make the endpoints
previously available under /transport/api available under
/api/v1/transport. Leave the old endpoints active for the time
being, but arrange for the auto-generated documentation to ignore them.
Redirect the old documentation location (/transport/api/docs) to the
new one (/api/docs).

There are two unsatisfactory things about the current approach:

1) /transport/api/urls is included twice (from /transport/urls and
/api/urls) but includes some name= parameters. Won't this cause
problems for Django's reverse() - which pattern will it choose?
However the alternative is to essentially duplicate a whole load of
url() calls
2) /api/urls has to define a bogus set of pattern to pass to
rest_framework.documentation.include_docs_urls() to get it to ignore
the URLs under /transport/api/urls, but not doing this and trying
to re-use the r'^v1/.../' patterns results in documentation that
links to the wrong URLs

Both of these will go away when the old endpoints can be retired.