Fix date logic for adding/removing subscriptions
Failure to correctly use .clone() when calculating the subscription window
caused the idea of 'now' to move forward by 30 minutes for each
journey processed, resulting in an incorrect set of subscriptions
and unsubscriptions being issued. This was largely benign, because
a bug elsewhere was causing subscriptions to be ignored so in most cases
we ended up with the sufficient active subscriptions.