Remove the child spec from sensor data serialisation
Raw sensor data is represented as a list of [time, reading] tupals.
Specifying this as a list of serializers.FloatField() causes the
time, which should be integer seconds, to be converted to a float.
It's not obvious how to specify a hetrogenious list (or even that it's
possible) so the pragmatic solution is to remove the specification. This
works, at the expense of the values not being validated.