Emergency fix to the Werther Widget backend
Today the Met Office are failing to include 'name' information
in the returned forecast data (at least fo Cambridge). This patch
adds a hard-coded station-id --> Name lookup table for the wether
stations that we are currently using and falls back to that if 'name'
is missing from the returned forecast.