Fall back to 'last known good' forecast on error
The Met Office API isn't all that reliable, in particular returning
data missing documented fields. Since each forecast is valid for multiple
days it makes sense to cache a 'last known good' version when we see one
and use that if whatever we are currently fetching is broken in any way.

Note that we still cache what we fetch (or an empty place holder if we
don't fetch anything) so even in an error condition we still only
hit the API once wvery 10 minuites for each forecast location requested.