Suppress journeys that terminate at the displayed stop
Without this fix, timetables may include journeys that just terminate
at the stop being displayed, which is unhelpful.

We could display these journeys as 'terminates here' which might be
useful if you were meeting a bus, but that's probably a minority
interest.

Worse, it's common for the vehicle involved in a terminating journey to
immediately move on to implement the next departing journey and will
probably be displaying the new journey's details on its headboard as it
arrives. For example, line 'U' journeys from Eddington terminate at
Addenbroke's Outpatients but the buses (almost always) immediately take
on the next journey back to Eddington and the drivers normally alter the
headboard before arriving. Because of this, passengers waiting at the
stop don't see a terminating journey, just the start of the next one. We
could do better if we knew if the vehicle was going out of service or
not, but currently that information isn't available.

Closes #169