Fix mismatch between zone transit plot and new API
1) Zone transit records in the new API don't include module_id,
or zone_id. Get zone_id for the point popup from the zone config
instead.

2) Zone transit records in the new API don't include the SIRI-VM
record that completed the zone. Leave his out if the point popup
(though log it as a potential enhancement)

3) Display points in green unless there's a route_id in the completion
data, in which case it comes from the GTFS feed handler and we
display it in green