Add ability to associate authorised referers to tokens
Tokens used by JavaScript applications are of necessity exposed.
To minimise the impact of this we follow Google's lead and allow
tokens to be associated with a list of referer URLs and reject access
using those tokens if the refers don't match. While this is easily
subverted in server-side scripts and by using custom proxies it does
make it slightly harder to misuse tokens exposed in this way.