Move general information out of auto-generated docs
It's not clear where general API information belongs, but there's
some convenience in separating the bulk of the auto-generated
documentation from manually-maintained stuff. This edit moves
revently written general information from the 'description'
section of the auto-generated documents back into the API's
homepage.

It also makes some editorial changes to the text, in particular
down-playing use of CoreAPI given that DRF apparently plans to move
from CoreAPI to OpenAPI in the next release.