Rework login flow so API docs no longer require authentication
It is difficult to test token-based authentication if you can't access
the API documentation without without being signed in. However it's
also difficult to arrange that people deliberately trying to use
session authentication will also have agreed to the TCs first.

This edit tries to get this right, by removing the authentication
requirement on /api/, adding one on the various token endpoints,
and providing an explicit authenticate and agree TCs link.