Change timetable/map display; 'panel' -> 'page'; default config
Rework the bus timetable and map so that the user can swap between them
without either being destroyed until they return to the list page. This
means that the real-time status and bus position information is
maintained. Note that the map is only initialised the first time it's
accessed (rather than at the same time as the timetable) because of the
difficulty of displaying a Leaflet map in a DOM element that isn't
currently displayed.

At the same time, reword both the user interface and the source code
to replace 'panel' (which has its own meaning elsewhere in the project)
as a way to describe 'widget instance' with 'page'. Unfortunately this
leads to two different use of 'page' in the code: as a 'widget instance'
and as an <ons-page> element. Where possible, the latter is referred to
as 'ons_page'.

At the same time, trigger the config process when an empty list
page is first displayed (as if the user had clicked '+').