Turn off bus_map tooltips and improve popup
Turn off tooltip messages on the stop_bus_map - they aren't relevant
on the main SmartPanels (which are non-interactive), and
Pocket Smartpanels are mainly displayed on tuch devices for which
hover (and hence tooltips) doesn't make sense.

At the same time, improve the popup messages by adding other things that
might be useful to a bus user.