Drop ES6 features from pocket.js
A significant number of Andriod browsers ship with a system browser
that can't cope with ES6. To help with this, this edit removes ES6
features from pocket.js, replacing them with equivalent ES5 features.

There is still a documented hard limit that Onsen UI requires Android
4.4.4 or above, though for many, installing/upgrading Andriod
Chrome may be a work-around for this.