Hide '+' while editing; rework show/hide
The swap from normal to edit mode is tricky to follow at the best of
times. To make it a bit more obvious, hide the '+' (add) icon when
in edit mode and change the page title.

At the same time, refactor hoe show/hide on entry/exit from edit mode
works to better handle lots of items..