Detect buses that should be sending real-time data but aren't
Buses that should have departed from their origin stop but which aren't
sending any real-time data are suspect - either their transmitter is bust,
or they are claiming to be on a different journey, or the trip has been
cancelled. Whatever is happening, it's useful to warn travellers that
this is happening. This edit adds the message 'Realtime data missing'
to any timetable entries for which this is the case.