Fix mis-alignment of the 'key' cell in timetables
Possibly as a result of changes to Chrome, the top and bottom borders
around the 'key' cell in the timetables rendered by stop_tmetable and
station_board had become mis-aligned. This was due to them honouring
a small shift applied to such cells to make their content
line up visually.

Fix this by applying to offset to a span within the cell, rather than to
the cell itself.