Disable 'Save' button unless a stop is chosen
It was previously possible to click 'Save' on the stop chooser page
without having chosen a stop. This resulted in an exception when
the resulting undefined stop reference was dereferenced.

This edit updates bus-stop-chooser to a version that supports a callback
that is called with the number of currently selected stops each time the
number of stops changes, and uses this to enable or disable the 'Save' button
as appropriate.

Closes #255