Switch to accessing WSDL files locally
The Pyhon Zeep SOAP library can, despite what the documentation implies,
access WSDL files directly via the file system. This edit re-adds the
cached files to the Djando project and adjusts the station_board
view to use them.

As well as being more efficient, this removes a dependency on the hosting
environment and makes it easier to run tfc_web from inside a Docker
container.

Closes #259