Move Onsen UI stylesheets above widget.css
widget.css needs to come second so that its style redefinitions
take precedence over those set by Onsen UI.

In particular, without this there's too much space above and below the
headings on the widget pages. this is particularly obvious on the map
page where the heading background doesn't touch the top of the display
area.