Rework instace_id and log it
Rework instance_id to be in the form 'AAAA-0000' and replace any
existing all-numeric ids.

Log instance_id to a dedicated logging endpoint each time the
app starts up.

Include instance_id as RTMonitor_api's 'Client ID' and as the
`display_id` passed to widget instances (not that they currently do
anything with this).