remove duplicate
smartcambridge_admin is not checking if the user is a superuser, is only checking that is_staff is checked which doesn't correspond to be an admin.

The same can be achieved used the altready provided decorator @user_passes_test(lambda u: u.is_superuser) so removing all the replicated and unused code.