Workaround for bug in everynet API
Everynet NS API for devices accepts upper and lower case hex as values for its input parameters but then internally applies lowercase to all of them when storing this values in everynet's database. Its delete device API endpoint has a bug because it doesn't apply the same filter (lowercase all inputs) into its input parameters and thus causes 404 not found when using an uppercase hex as an input. This PR forces all webforms in CSN to store lowercase for all hex values in our database to avoid facing this bug in any of our API calls.