Rework station_board view to better cope with errors
As with weather, arrange to cache the last known good station board
for a particular station and use that data if an attempt to retrieve
new information fails. As well as allowing station board widgets to
ride over outages of the underlying service, this will prevent
exceptions being thrown when this happens and so avoid torrents
of emails from the logging system.

Unlike weather, only cache last known good data for half an hour
since it will be stale after that.