Avoid error when we have multiple entries with the same slug
This doesn't solve the problem but it avoids throwing an error. We need to make slug to be unique in the data model to avoid this occurring as the we won't be able to access to all those entries that have the same slug. This is an edge case, though.