Restrict emailing errors to production
Move the definition of the logger that causes emails to be sent
on errors from settings.py (which is used by everything) to
settings_production.py (which is used in production).

The intension is to stop emails being generated on development
machines (which are probably running with DEBUG=True and so will
show tracebacks anyway).