Fix capacity annotation on car park graphs
The capacity marker (red line and '(max xxxx)' wording) for the car
parks comes from the static car park metadata which can't cope with
changes to capacity over time.

By comparison, the list and the map pages get this information comes
from the 'capacity' field of the occupancy record.

This edit changes the graph logic to use the capacity from the last
occupancy record for the day being plotted. This still isn't quite right
for a day on which the capacity changes or given decreeing capacity and
prior_days plots, but since these are rare events it's not worth
worrying about.

Closes #319