Restore adding default pages to a new instance
Restore the logic that adds Cambridge station and Cambridge weather
(but not any bus stops) to a new instance of the panel. Necessarily
rework the logic that displays the 'first time' usage hint.

Be careful only to save the list of pages (added by default or from a
preload list) to local storage after the user has consented to our
doing so.