Allow stop code (ATCOCode or NAPTAN code) in URL
Add to the existing Pocket SmartPanel preload system the ability to
specify a single bus stop, by ATCOCode or NaPTAN Code, in a URL. This
stop will be added to the Pocket SmartPanel configuration.

The expected use case for this involves provisioning selected bus stops
with notices including a custom URL and/or QR Code that will add that
stop to people's configuration.

This is just PoC code and needs more work for production. In particular
the pages it creates don't have the same title or set of destinations
that a page created interactively would have. Fixing this either
requires duplicating Pocket SmartPanel javaScript logic in the Stops
model and/or the Pocket SmartPanel view, or re-factoring to make this
unnecessary.