Fixup initial save of preloaded pages
The list of current pages was previously saved to local storage
when the OK button was pressed on the first.html page. This fails to
capture pages loaded through the 'preload' mechanism when an existing
instance is launched. It's not safe to save preloaded pages during
initial processing following the 'page ready' event because the user
may not yet have given consent.

This edit adds a page save at the top if the initialisation code for
the list page. In almost all cases this will just re-save the state
that was just loaded, but it does ensure that preloads are handled
correctly. In practice this is only run once on each initialisation
so it won't be hugely inefficient.