Implement rate limiting for traffic-map
Google charge us actual money for every load of a traffic map, though they
give us a useful free usage allocation each month. SmartPanels only do
a chargeable reload when they start up, and on our recommended RaspberryPI
platforms this normally only happens as a result of a reboot once per day. At this
rate, our free allocation can support a large number of SmartPanels at
no cost.

However it is possible to deploy SmartPanels in other ways that result
in more frequent reloads. For example as one page in an 'carousel'
displaying a sequence of useful pages. If this results in a reload
every minute this takes us from 1 chargeable load per day to 1,440 loads
per day. This quickly exceeds our free allocation.

This edit addresses this as follows:

1) All loads of the Google map by the traffic-map widget are tracked in
local storage on each device displaying a SmartPanel. Usage above some
threshold blocks further loads of the map. The default per-day limit is
set by the SMARTPANEL_TRAFFIC_MAP_RELOAD_LIMIT_DEFAULT key in settings.py
(currently 10 loads per day).

Successful and failing loads are logged to the
same endpoint used for logging Pocket SmartPanel usage under the
module_id 'smartpanel'.

Use of local storage means this isn't a secure
way of restricting use (because it's easy enough to delete local storage)
but it is sufficient to prevent mistakes.

2) Displays have a new parameter map_reload_limit which, if set,
overrides the default limit of 10 loads per day for that particular
display. There's no UI for setting this, other than Django admin.

3) Each traffic-map instance has a new configuration parameter that can
contain a Google Maps API key. If present, this key is used in place of
the default SmartCambridge one. This way, any cost associated with
the map falls on the person configuring the key. If a custom key is
present then no rate limiting is applied.

The Google API key is supplied as part of the URL that loads the
Googe JavaScript library. To achieve this, the Google Maps API URL
has been taken out of requirements.json and is now generated on the
fly in the code that sets up all the other variables for the
layout.html template. The code actually uses the first custom API key
found - on a display with multiple traffic-maps it is undefined which of
several different keys would actually be used.