Fix login-and-agree
The login-and-agree view supports an optional 'next=' parameter which
can supply a URL to go to following successful authentication. This
has never worked, because supplying settings.ALLOWED_HOSTS as the value
of allowed_hosts in a call to django.utils.http.is_safe_url() doesn't
actually work (is_safe_url doesn't handle values of '*' or domain name
suffixes).

This edit uses the value of request.get_host() as the (single) value for
allowed_hosts. While this comes from the Host: header and so isn't to be
trusted, I think this is OK in this case because it's not under the
control of anyone likely to be trying to exploit the redirect behaviour.
In any case this seems to be what the default Django login view does
under the same circumstances.