Let us cling on to using CoreAPI
Django REST Framework is deprecating CoreAPI based schema generation in
favour of OpenAPI schema. In theory that's a good idea, if it wasn't for
the investment we have in CoreAPI documentation:

https://github.com/encode/django-rest-framework/issues/6809

See https://www.django-rest-framework.org/community/3.10-announcement/#continuing-to-use-coreapi

This edit adds 'DEFAULT_SCHEMA_CLASS' to REST_FRAMEWORK settings to preserve
use of CoreAPI for the time being, and pins djangorestframework < v3.12
where the remaining CoreAPI functionality is due to be withdrawn.

https://github.com/SmartCambridge/tfc_web/issues/346 notes the need to
upgrade.