Fix the menu on the Pocket SmartPanel usage graphs
The graph pages were inheriting from base.html, rather then
smartpanel/base.html and so had the main site left-hand menu rather then
the one used on other SmartPanel pages. This was particularly silly because
the SmartPanel menu is the only place that links to the graphs appear.
This edit fixes this.