Disable automatic generation of bus data archives
This reverts a change in c9d26996. This is because processing bus archives
is expensive because of the number of files involved and the need to
check each DTM to see if archives need to be regenerated. Combined
with the fact that bus archives are only monthly this would make for
a lot of wasted work on all but one day in any month.

This edit disables automatic generation of bus data archives
(`'archive_by_default': False,`), with the intention of running a
separate cron job on the first of every month to deal with buses.