Add journey details to API responses
Internally, zone completion messages include a 'position_record' field
containing a copy of the SIRI-VM record for the position that triggered
zone completion. The old 'dataserver' api included this in the data it
returned but the REST API didn't. This data has some utility (it supplies
data about the service and timetable journey, for example, and it was
displayed in a popup on the zone transit graph).

This edit adds the useful bits of the SIRI-VM record (the ones relating
to the timetabled journey) into the data returned by the API.

[If I'd known how easy this would be to do I'd have done it sooner!]

Closes #203