Add (restore) journey details to popup
Before the move to the REST API, the popup on the journey time graph
showed the SIRI-VM record that caused the zone completion. This
functionality was lost in the move because the REST API didn't include
this data.

Following #dd62599 which added journey data to the API response, this
edit restores journey details to the popup.

It also slightly reformats the data in the popup and adds units to
to description of the most recent point on the graph.