Rework download API configuration and download index page
Rework the download configuration to support more than one metadata
file per feed, and to support building metadata files from multiple
source files.

As part of this, rework generation of the the download index page
to rely more on information in the download API configuration and less
on prior knowledge of filesystem layout.