Add BT Journey visualisations
Add a map and a list of BT Journey point-to-point links and compound
routes, both linking through to plots of journey time. The map additionally
shows live journey time information. All data comes from the BT Journey API.
Apart from the live journey time display, this is all very similar to the
zone transit data visualisations.

The map is heavily based on the one developed in
https://github.com/SmartCambridge/datafeeds/tree/master/drakewell/visualise
but with significant internal reworking to cope with differing behaviour
between the Drakewell API and ours. The list and the plots are based on
the code used for zones.

The map loads OSM tiles over HTTPS to address new cross-site resource
warnings in Chrome.

This edit removes information about the link CAMBRIDGE_JTMS|9800WBETRSU3
from the test data to simulate the situation in which journeytimes data
references a link (or route) which is not present in locations (which happens
for a short while whan a new link or route is created.