Get rid of api/api_docs.py
This module used to include various coreapi field definitions that
were in theory used throughout the API. In practice only one definition
was actually used in more than one place, and having these definitions
so far from where they were used was unhelpful.

This edit moves the definitions into the files that use them (duplicating
the one that was used in two places) and deletes the now empty file.