Work around period appearing as {} (empty object)
For reasons unknown, the period field in journey data is occasionally
an empty object rather than an integer. In particular this has been
observed for very recently created links
(see e.g. CAMBRIDGE_JTMS|9800YRAA8RIZ on 2020-02-16)

This causes the corresponding searilizer to complain. This fix overwrites
such valuse with None which searalizes as 'null' as expected.