Also work around period == {} for download API
As mentioned elsewhere, for reasons unknown, the period field in journey
data is occasionally an empty object rather than an integer. In particular
this has been observed for very recently created links
(see e.g. CAMBRIDGE_JTMS|9800YRAA8RIZ on 2020-02-16)

The download API ends up copying the stringification of the empty object
('{}') into the CSV file which will probably be a surprise to anyone using
it. This edit overwrites such values with None which are searalized
as an empty field as expected.