Also work around period = {} for download API
As mentioned elsewhere, for reasons unknown the period field in journey
data is occasionally an empty object rather than an integer. In particular
this has been observed for very recently created links (see e.g.
CAMBRIDGE_JTMS|9800YRAA8RIZ on 2020-02-16)

The download API previously included '{}' in the CSV which is likely to
surprise people using it. This fix overwrites such values with None which
generates a blank field in the CSV which, even though not quite right is
the best we can achieve.