Always un-hide the config area before triggering widget .config methods
Prior to this change, when editing an existing widget the widget config area
was hidden while being populated by the widget's .config method and only
subsequently 'un-hidden'. This wasn't the case for a new widget because the
widget config area had to be displayed to reveal the dropdown so a type
could be selected.

Unfortunately, Leaflet maps can't be instantiated into DOM elements that
are hidden (or which aren't added to the document instance) because
Leaflet sees such elements as having zero size, and some widget config
methods display maps which therefore failed to display the second and
subsequent time the widget was edited.

This edit moves the line that displays the widget config area to before
the code that triggers both choosing new widgets and editing existing
ones.

This bug affected the (un-deployed) bike share widget and and in-development
BT Journey widget.